Create numberGuessingGame.java
In this game, the user should guess numbers 0 or 1. If the user is correct for the first time then the user gets 10 Points and the user moves to the next Stage. In the next stage, points are increased by 10, and two numbers are increased by one i.e. numbers will be 0,1, and 2. The pattern of the score for each gameplay is 10,20,30,40,...